+++ fst/layout/PlainLayout.cc
@@ -33,7 +33,7 @@ EOSFSTNAMESPACE_BEGIN
 // Constructor
 //------------------------------------------------------------------------------
 PlainLayout::PlainLayout (XrdFstOfsFile* file,
-                          int lid,
+                          unsigned long lid,
                           const XrdSecEntity* client,
                           XrdOucErrInfo* outError,
                           eos::common::LayoutId::eIoType io,
