+++ fst/layout/ReedSLayout.cc
@@ -43,7 +43,7 @@ EOSFSTNAMESPACE_BEGIN
 // Constructor
 //------------------------------------------------------------------------------
 ReedSLayout::ReedSLayout(XrdFstOfsFile* file,
-                         int lid,
+                         unsigned long lid,
                          const XrdSecEntity* client,
                          XrdOucErrInfo* outError,
                          eos::common::LayoutId::eIoType io,
