+++ fst/layout/ReplicaParLayout.hh
@@ -51,7 +51,7 @@ public:
   //!
   //--------------------------------------------------------------------------
   ReplicaParLayout (XrdFstOfsFile* file,
-                    int lid,
+                    unsigned long lid,
                     const XrdSecEntity* client,
                     XrdOucErrInfo* outError,
                     eos::common::LayoutId::eIoType io,
